# backend/docs/swagger.yaml
basePath: /
definitions:
...
  entity.BatchDeleteLoginLogDto:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
  entity.BatchDeleteOpLogDto:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
  entity.BatchDeletePostsDto:
    properties:
      postIds:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - postIds
    type: object
...
  entity.CreateSysPostDto:
    properties:
      postCode:
        type: string
      postName:
        type: string
      psotStatus:
        enum:
        - 1
        - 2
        type: integer
      remark:
        type: string
    required:
    - postCode
    - postName
    type: object
...
  entity.DeleteLoginLogDto:
    properties:
      id:
        type: integer
    type: object
...
  entity.DeleteOpLogDto:
    properties:
      id:
        type: integer
    type: object
  entity.DeletePostByIdDto:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
...
  entity.GetPostByIdDto:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
...
  entity.UpdatePostStatusDto:
    properties:
      id:
        type: integer
      newStatus:
        enum:
        - 1
        - 2
        type: integer
    required:
    - id
    type: object
...
  entity.UpdateSysPostDto:
    properties:
      id:
        type: integer
      postCode:
        type: string
      postName:
        type: string
      postStatus:
        enum:
        - 1
        - 2
        type: integer
      remark:
        type: string
    required:
    - id
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 后台管理系统API接口文档
  title: go-admin 后台管理系统
  version: "1.0"
paths:
...
  /api/logService/batchDeleteLoginLog:
    post:
      consumes:
      - application/json
      description: 批量删除登录日志
      parameters:
      - description: 批量删除登录日志请求
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.BatchDeleteLoginLogDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 批量删除登录日志
      tags:
      - 日志管理
  /api/logService/batchDeleteOpLog:
    post:
      consumes:
      - application/json
      description: 批量操作登录日志
      parameters:
      - description: 批量删除操作日志请求
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.BatchDeleteOpLogDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 批量删除操作日志
      tags:
      - 日志管理
  /api/logService/deleteLoginLog:
    post:
      consumes:
      - application/json
      description: 删除登录日志
      parameters:
      - description: 删除登录日志请求
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.DeleteLoginLogDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除登录日志
      tags:
      - 日志管理
  /api/logService/deleteOpLog:
    post:
      consumes:
      - application/json
      description: 删除操作日志
      parameters:
      - description: 删除操作日志请求
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.DeleteOpLogDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除操作日志
      tags:
      - 日志管理
  /api/logService/getLoginLogList:
    get:
      consumes:
      - application/json
      description: 查询登录日志列表
      parameters:
      - description: 页码
        in: query
        name: pageNum
        type: integer
      - description: 页大小
        in: query
        name: pageSize
        type: integer
      - description: 用户名
        in: query
        name: username
        type: string
      - description: 登录状态
        in: query
        name: loginStatus
        type: integer
      - description: 开始时间
        in: query
        name: beginTime
        type: string
      - description: 结束时间
        in: query
        name: endTime
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 查询登录日志列表
      tags:
      - 日志管理
  /api/logService/getOperationLogList:
    get:
      consumes:
      - application/json
      description: 查询操作日志列表
      parameters:
      - description: 页码
        in: query
        name: pageNum
        type: integer
      - description: 页大小
        in: query
        name: pageSize
        type: integer
      - description: 用户名
        in: query
        name: username
        type: string
      - description: 开始时间
        in: query
        name: beginTime
        type: string
      - description: 结束时间
        in: query
        name: endTime
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 查询操作日志列表
      tags:
      - 日志管理
...
  /api/postService/batchDeletePosts:
    post:
      consumes:
      - application/json
      description: 批量删除岗位
      parameters:
      - description: 要删除的岗位id列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.BatchDeletePostsDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 批量删除岗位
      tags:
      - 岗位管理
  /api/postService/createPost:
    post:
      consumes:
      - application/json
      description: 创建岗位
      parameters:
      - description: 创建岗位的请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.CreateSysPostDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建岗位
      tags:
      - 岗位管理
  /api/postService/deletePost:
    post:
      consumes:
      - application/json
      description: 删除单个岗位
      parameters:
      - description: 删除单个岗位请求结构体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.DeletePostByIdDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除单个岗位
      tags:
      - 岗位管理
  /api/postService/getPostById:
    post:
      consumes:
      - application/json
      description: 根据id查询岗位信息
      parameters:
      - description: 获取岗位信息请求结构体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.GetPostByIdDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 根据id查询岗位信息
      tags:
      - 岗位管理
  /api/postService/getPostDropdown:
    get:
      consumes:
      - application/json
      description: 岗位下拉列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 岗位下拉列表
      tags:
      - 岗位管理
  /api/postService/getPostList:
    get:
      consumes:
      - application/json
      description: 分页获取岗位列表
      parameters:
      - description: 页码
        in: query
        name: pageNum
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 岗位状态：1->启用,2->禁用
        in: query
        name: postStatus
        type: integer
      - description: 岗位名称
        in: query
        name: postName
        type: string
      - description: 开始时间
        in: query
        name: beginTime
        type: string
      - description: 结束时间
        in: query
        name: endTime
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取岗位列表
      tags:
      - 岗位管理
  /api/postService/updatePost:
    post:
      consumes:
      - application/json
      description: 修改岗位信息
      parameters:
      - description: 修改岗位请求结构体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateSysPostDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 修改岗位
      tags:
      - 岗位管理
  /api/postService/updatePostStatus:
    post:
      consumes:
      - application/json
      description: 修改岗位状态
      parameters:
      - description: 修改岗位状态请求
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.UpdatePostStatusDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 修改岗位状态
      tags:
      - 岗位管理
...
securityDefinitions:
  BearerAuth:
    description: JWT认证令牌，格式Bearer <token>
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"